requirements.txt
manim==0.18.1
Pillow==10.4.0
